Creating LawNumber12982 style file
