+++ src/Styles/LegislationStyles/LawNumber12982Styles.js
@@ -1,0 +1,31 @@
+import { StyleSheet } from 'react-native';
+
+
+const styles = StyleSheet.create({
+    title: {
+      textAlign: 'center',
+      fontSize: 22,
+      fontWeight: '300',
+      marginBottom: 20,
+    },
+    selectors: {
+      marginBottom: 10,
+      flexDirection: 'row',
+      justifyContent: 'center',
+    },
+    selector: {
+      backgroundColor: '#F5FCFF',
+      padding: 10,
+    },
+    activeSelector: {
+      fontWeight: 'bold',
+    },
+    selectTitle: {
+      fontSize: 14,
+      fontWeight: '500',
+      padding: 10,
+    },
+});
+
+export default styles;
+
